+++ mission1/src/Components/ShoppingList.css
@@ -1,0 +1,39 @@
+.shoppingList {
+    list-style: none;
+    width: 98%
+  }
+
+  ul {
+    margin: 0;
+    padding: 0;
+  }
+
+  img {
+    object-fit: cover;
+    width: 30px;
+    height: 30px;
+  }
+  
+  ul li {
+    cursor: pointer;
+    position: relative;
+    padding: 12px 8px 12px 40px;
+    background: #eee;
+    font-size: 18px;
+    transition: 0.2s;
+  
+    /* make the list items unselectable */
+    -webkit-user-select: none;
+    -moz-user-select: none;
+    -ms-user-select: none;
+    user-select: none;
+  }
+  
+  ul li:nth-child(odd) {
+    background: #f9f9f9;
+  }
+  
+  /* Darker background-color on hover */
+  ul li:hover {
+    background: #ddd;
+  }
