+++ mission1/src/style.css
@@ -16,71 +16,3 @@ body {
   width: 100%;
   overflow: auto;
 }
-
-.shoppingList {
-  list-style: none;
-  width: 98%
-}
-
-.ListItem {
-  display: flexbox;
-  float: left;
-  margin-right: 20px;
-  margin-left: 20px;
-  width: 300px;
-  height: 100px;
-  margin-top: 15px;
-  border: 1px solid black;
-  box-sizing: border-box;
-  font-family: "Syne Mono", monospace;
-  border-radius: 10px;
-}
-
-.ListItem__name {
-  font-weight: 1000;
-}
-
-.ListItem__img {
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  left: 60%;
-  object-fit: cover;
-  width: 60px;
-  height: 60px;
-}
-
-ul {
-  margin: 0;
-  padding: 0;
-}
-
-img {
-  object-fit: cover;
-  width: 30px;
-  height: 30px;
-}
-
-ul li {
-  cursor: pointer;
-  position: relative;
-  padding: 12px 8px 12px 40px;
-  background: #eee;
-  font-size: 18px;
-  transition: 0.2s;
-
-  /* make the list items unselectable */
-  -webkit-user-select: none;
-  -moz-user-select: none;
-  -ms-user-select: none;
-  user-select: none;
-}
-
-ul li:nth-child(odd) {
-  background: #f9f9f9;
-}
-
-/* Darker background-color on hover */
-ul li:hover {
-  background: #ddd;
-}
